# swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/workspaces/{workspaceId}/pages": {
      "post": {
        "operationId": "PagesController_create",
        "summary": "Create a new page in workspace",
        "parameters": [
          {
            "name": "workspaceId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreatePageDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Page created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PageResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          }
        },
        "tags": [
          "Pages"
        ],
        "security": [
          {
            "JWT-auth": []
          }
        ]
      },
      "get": {
        "operationId": "PagesController_getTree",
        "summary": "Get page tree for workspace",
        "parameters": [
          {
            "name": "workspaceId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Page tree",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PageTreeResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          }
        },
        "tags": [
          "Pages"
        ],
        "security": [
          {
            "JWT-auth": []
          }
        ]
      }
    },
    "/pages/{id}": {
      "get": {
        "operationId": "PagesController_findOne",
        "summary": "Get page by ID",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Page details",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PageResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Page not found"
          }
        },
        "tags": [
          "Pages"
        ],
        "security": [
          {
            "JWT-auth": []
          }
        ]
      },
      "patch": {
        "operationId": "PagesController_update",
        "summary": "Update page",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdatePageDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Page updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PageResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Page not found"
          }
        },
        "tags": [
          "Pages"
        ],
        "security": [
          {
            "JWT-auth": []
          }
        ]
      },
      "delete": {
        "operationId": "PagesController_delete",
        "summary": "Delete page (soft delete)",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Page deleted successfully"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Page not found"
          }
        },
        "tags": [
          "Pages"
        ],
        "security": [
          {
            "JWT-auth": []
          }
        ]
      }
    },
    "/pages/{id}/move": {
      "post": {
        "operationId": "PagesController_move",
        "summary": "Move page to new parent or position",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MovePageDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Page moved successfully"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Page not found"
          }
        },
        "tags": [
          "Pages"
        ],
        "security": [
          {
            "JWT-auth": []
          }
        ]
      }
    },
...
    "/reference/organizations/{id}/tree": {
      "get": {
        "operationId": "OrganizationsController_getTree",
        "summary": "Get organization tree",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Organizations"
        ]
      }
    },
    "/reference/organizations/{id}/hierarchy": {
      "get": {
        "operationId": "OrganizationsController_getHierarchy",
        "summary": "Get organization hierarchy (flat list)",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Organizations"
        ]
      }
    },
    "/reference/organizations/{id}": {
      "get": {
        "operationId": "OrganizationsController_findById",
        "summary": "Get organization by ID",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Returns organization",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrganizationResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "Organization not found"
          }
        },
        "tags": [
          "Organizations"
        ]
      }
    },
...
  },
  "info": {
    "title": "Elbruso API",
    "description": "BI System REST API Documentation",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "JWT-auth": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "CreatePageDto": {
        "type": "object",
        "properties": {
          "parentPageId": {
            "type": "string",
            "example": "550e8400-e29b-41d4-a716-446655440000"
          },
          "title": {
            "type": "string",
            "example": "My Page"
          },
          "icon": {
            "type": "string",
            "example": "📄"
          },
          "coverImage": {
            "type": "string",
            "example": "https://example.com/cover.jpg"
          }
        },
        "required": [
          "title"
        ]
      },
      "PageResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "title": {
            "type": "string"
          },
          "content": {
            "type": "string"
          },
          "icon": {
            "type": "string"
          },
          "coverImageUrl": {
            "type": "string"
          },
          "isPublished": {
            "type": "boolean"
          },
          "isFavorite": {
            "type": "boolean"
          },
          "parentId": {
            "type": "string",
            "format": "uuid"
          },
          "groupId": {
            "type": "string",
            "format": "uuid"
          },
          "workspaceId": {
            "type": "string",
            "format": "uuid"
          },
          "order": {
            "type": "number"
          },
          "createdBy": {
            "type": "string"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "id",
          "title",
          "content",
          "icon",
          "coverImageUrl",
          "isPublished",
          "isFavorite",
          "parentId",
          "groupId",
          "workspaceId",
          "order",
          "createdBy",
          "createdAt",
          "updatedAt"
        ]
      },
      "PageTreeItemResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "title": {
            "type": "string"
          },
          "icon": {
            "type": "string"
          },
          "order": {
            "type": "number"
          },
          "isPublished": {
            "type": "boolean"
          },
          "children": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PageTreeItemResponseDto"
            }
          }
        },
        "required": [
          "id",
          "title",
          "icon",
          "order",
          "isPublished",
          "children"
        ]
      },
      "PageTreeResponseDto": {
        "type": "object",
        "properties": {
          "tree": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PageTreeItemResponseDto"
            }
          },
          "total": {
            "type": "number"
          }
        },
        "required": [
          "tree",
          "total"
        ]
      },
      "UpdatePageDto": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string",
            "example": "Updated Title"
          },
          "icon": {
            "type": "string",
            "example": "📈"
          },
          "coverImage": {
            "type": "string",
            "example": "https://example.com/new-cover.jpg"
          }
        }
      },
      "MovePageDto": {
        "type": "object",
        "properties": {
          "parentPageId": {
            "type": "string",
            "example": "550e8400-e29b-41d4-a716-446655440002",
            "description": "New parent page ID. Null for root level."
          },
          "afterPageId": {
            "type": "string",
            "example": "550e8400-e29b-41d4-a716-446655440003",
            "description": "Insert after this page. Null for first position."
          }
        }
      },
...
      "OrganizationResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number",
            "description": "Organization ID"
          },
          "name_ru": {
            "type": "string",
            "description": "Organization name in Russian"
          },
          "abbreviation_ru": {
            "type": "string",
            "description": "Abbreviation in Russian",
            "nullable": true
          },
          "internal_code": {
            "type": "string",
            "description": "Internal code",
            "nullable": true
          },
          "type_id": {
            "type": "number",
            "description": "Organization type ID"
          },
          "parent_id": {
            "type": "number",
            "description": "Parent organization ID",
            "nullable": true
          },
          "sport_id": {
            "type": "number",
            "description": "Sport ID",
            "nullable": true
          },
          "region_id": {
            "type": "number",
            "description": "Region ID",
            "nullable": true
          },
          "country_id": {
            "type": "number",
            "description": "Country ID",
            "nullable": true
          },
          "level_id": {
            "type": "number",
            "description": "Organization level ID"
          },
          "founded_year": {
            "type": "number",
            "description": "Founded year",
            "nullable": true
          },
          "is_active": {
            "type": "boolean",
            "description": "Active status"
          },
          "metadata": {
            "type": "object",
            "description": "Metadata (JSON)",
            "nullable": true
          },
          "created_at": {
            "format": "date-time",
            "type": "string",
            "description": "Creation timestamp"
          },
          "updated_at": {
            "format": "date-time",
            "type": "string",
            "description": "Last update timestamp"
          }
        },
        "required": [
          "id",
          "name_ru",
          "type_id",
          "level_id",
          "is_active",
          "created_at",
          "updated_at"
        ]
      },
...
    }
  }
}